requirements.txt
Flask ==3.0.2 
python-dotenv
pymongo[srv]
# gunicorn


